requirements.txt
streamlit
pillow
matplotlib